Move encoding out from keyring condition
Since b_wincred_set_with_raw_value used to use `charToRaw` on the password from within the argument, the password was always raw here; passing it as text broke the functionality. 

Moving it back into the main body of the function -- thereby encoding the password regardless of keyring in use -- fixes this problem.